+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## [Unreleased]
 
+## [0.6.10] - 2023-08-01
+- 🗣️ LLMs
+  - Introducing Anthropic support
+
+## [0.6.9] - 2023-07-29
+
+## [0.6.8] - 2023-07-21
+
+## [0.6.7] - 2023-07-19
+- Support for OpenAI functions
+- Streaming vectorsearch ask() responses
+
 ## [0.6.6] - 2023-07-13
 - Langchain::Chunker::RecursiveText
 - Fixes